+++ Tests/AccessibilityTests/AccessibilityProject/Sources/TargetA/PublicComparableOperatorFunction.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+public class PublicComparableOperatorFunction: Comparable {}
+
+public func < (lhs: PublicComparableOperatorFunction, rhs: PublicComparableOperatorFunction) -> Bool {
+    true
+}
+
+public func == (lhs: PublicComparableOperatorFunction, rhs: PublicComparableOperatorFunction) -> Bool {
+    true
+}